トップアクセス時、slugが設定されたurlにリダイレクトするよう調整
